Fix int/long issue

+++ src/com/scurab/java/ftpleechergui/model/Settings.java
@@ -13,7 +13,7 @@ public class Settings {
      * Default value is 15MB
      */
 
-    public int globalPieceLength = 15000000;
+    public long globalPieceLength = 15000000;
 
     /**
      * Allow resuming of downloading parts.<br/>
